improvements + separated patch from patches
