+++ src/Patches.cs
@@ -16,205 +16,29 @@ using UndertaleModLib;
 using UndertaleModLib.Models;
 using UndertaleModLib.Scripting;
 using static System.Runtime.InteropServices.JavaScript.JSType;
+using static UndertaleModLib.Compiler.Compiler.AssemblyWriter;
 
 namespace cpatcher;
 
-/// <summary>
-/// exception what is thrown during an error in execution of a patch
-/// </summary>
-public class PatchException : Exception
-{
-    public PatchException() { }
-    public PatchException(string message) : base(message) { }
-}
-
-public class PatchGlobals
-{
-    /// <summary>
-    /// the circloo data win
-    /// </summary>
-    public UndertaleData Data;
-
-    /// <summary>
-    /// path to the root of circloo
-    /// </summary>
-    public string CircloORootPath { get; init; }
-
-    /// <summary>
-    /// path to the currently executed patch
-    /// </summary>
-    public string PatchPath { get; init; }
-
-    /// <summary>
-    /// Info about current patch
-    /// </summary>
-    public PatchInfo Info { get; init; }
-
-    // WIP: some actual error handling for patches, assembler (asm) integration
-
-    // source: cirqol
-    /// <summary>
-    /// get undertale code inside data from code name, and create it if doesnt exist, as an option
-    /// </summary>
-    public UndertaleCode? Code(string codeName, bool makeIfNotExists = true)
-    {
-        UndertaleCode code = Data.Code.ByName(codeName);
-        if (code == null && makeIfNotExists)
-        {
-            UndertaleCodeLocals locals = new UndertaleCodeLocals();
-            locals.Name = Data.Strings.MakeString(codeName);
-            code = new UndertaleCode();
-            code.Name = locals.Name;
-            Data.Code.Add(code);
-            Data.CodeLocals.Add(locals);
-        }
-        return code;
-    }
-
-    /// <summary>
-    /// create gml function from gml code
-    /// </summary>
-    public void CreateGMLFunction(string code)
-    {
-        string funcname = Regex.Match(code, @"^\s*function\s*(\w+)").Groups[1].Value;
-        string filename = "gml_GlobalScript_" + funcname;
-        Code(filename).ReplaceGML(code, Data);
-    }
-
-    /// <summary>
-    /// show some message (information) to the user
-    /// </summary>
-    public void PatchMessage(string message)
-    {
-        MessageBox.Show(message, $"Message from {Info.DisplayName}", MessageBoxButtons.OK, MessageBoxIcon.Information);
-    }
-
-    /// <summary>
-    /// show some error to the user
-    /// </summary>
-    public void PatchError(string error)
-    {
-        MessageBox.Show(error, $"Error from {Info.DisplayName}", MessageBoxButtons.OK, MessageBoxIcon.Error);
-    }
-
-    /// <summary>
-    /// ask a yes/no question to the user
-    /// </summary>
-    public bool PatchQuestion(string message)
-    {
-        DialogResult res = MessageBox.Show(message, $"Question from {Info.DisplayName}", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-        return res == DialogResult.Yes;
-    }
-
-    public PatchGlobals(UndertaleData data, PatchInfo info, string rootPath, string patchPath)
-    {
-        Data = data;
-        Info = info;
-        CircloORootPath = rootPath;
-        PatchPath = patchPath;
-    }
-}
-
-public readonly struct PatchInfo
-{
-    public string Name { get; init; } // aka internal name, init makes it so it cant be changes afterwards
-    public string DisplayName { get; init; }
-    public string Description { get; init; }
-    public string Version { get; init; }
-    public string Author { get; init; }
-    public PatchInfo(string name, string displayName, string description, string version, string author)
-    {
-        Name = name;
-        DisplayName = displayName;
-        Description = description;
-        Version = version;
-        Author = author;
-    }
-}
-
-public class Patch
-{
-    public Guid Id { get; init; }
-    public PatchInfo Info { get; init; }
-    public string CodePath { get; set; }
-
-    public Patch(string codePath, PatchInfo info)
-    {
-        Id = Guid.NewGuid();
-        CodePath = codePath;
-        Info = info;
-    }
-
-    // samples:
-    // https://github.com/dotnet/roslyn/blob/main/docs/wiki/Scripting-API-Samples.md
-    // https://github.com/UnderminersTeam/UndertaleModTool/blob/fb312d9a2b64063dcd458b9cdec3acb06c081db0/UndertaleModTool/ScriptingFunctions.cs#L58
-    // https://github.com/UnderminersTeam/UndertaleModTool/blob/fb312d9a2b64063dcd458b9cdec3acb06c081db0/UndertaleModTool/MainWindow.xaml.cs#L248
-    public bool Execute()
-    {
-        if (!File.Exists(CodePath))
-        {
-            return false;
-        }
-
-        try
-        {
-            ScriptOptions scriptOptions = ScriptOptions.Default
-                .AddImports("UndertaleModLib", "UndertaleModLib.Models", "UndertaleModLib.Decompiler",
-                    "UndertaleModLib.Scripting", "UndertaleModLib.Compiler",
-                    "UndertaleModTool", "System", "System.IO", "System.Collections.Generic",
-                    "System.Text.RegularExpressions", "System.Linq")
-                .AddReferences(typeof(UndertaleObject).GetTypeInfo().Assembly, typeof(System.Text.RegularExpressions.Regex).GetTypeInfo().Assembly)
-                .WithEmitDebugInformation(true); // when script throws an exception, add a exception location (line number)
-
-            //ScriptOptions scriptOptions = ScriptOptions.Default;
-
-            CancellationTokenSource source = new CancellationTokenSource(100);
-            CancellationToken token = source.Token;
-
-            PatchGlobals globals = new PatchGlobals(MainWindow.Data, Info, MainWindow.CircloORootPath, CodePath);
-
-            object result = CSharpScript.EvaluateAsync(
-                File.ReadAllText(CodePath),
-                scriptOptions,
-                globals,
-                typeof(PatchGlobals),
-                token
-            );
-
-            MainWindow.Data = globals.Data;
-        }
-        catch (CompilationErrorException exc)
-        {
-            MessageBox.Show($"Compilation error of patch {Info.DisplayName}: {exc}", $"Error from {Info.DisplayName}", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            return false;
-        }
-        catch (Exception)
-        {
-            return true;
-        }
-
-        return true;
-    }
-}
-
 public partial class MainWindow
 {
-    public List<Patch> Patches { get; set; } = new List<Patch>();
-    public List<Patch> SelectedPatches { get; set; } = new List<Patch>();
+    public List<Patch> Patches { get; set; } = new();
+    public List<Patch> SelectedPatches { get; set; } = new();
 
-    private Patch? FindByName(string name)
+    private bool AnyPatch(Patch patch, List<Patch> listPatch)
     {
-        return Patches.Find(p => p.Info.Name == name);
+        return listPatch.Any(p => p.Info.Name == patch.Info.Name);
     }
+
     public List<Patch> GetSelectedPatches()
     {
-        List<Patch> selectedPatches = new List<Patch>();
+        List<Patch> selectedPatches = new();
         if (Properties.Settings.Default.SelectedPatches != null)
         {
-            foreach (string patchName in Properties.Settings.Default.SelectedPatches)
+            foreach (string? patchName in Properties.Settings.Default.SelectedPatches)
             {
-                Patch? foundPatch = FindByName(patchName);
-                if (foundPatch != null)
+                Patch? foundPatch = Patches.Find(p => p.Info.Name == patchName);
+                if (foundPatch != null && !AnyPatch(foundPatch, selectedPatches))
                 {
                     selectedPatches.Add(foundPatch);
                 }
@@ -292,12 +116,13 @@ public partial class MainWindow
 
     public void NewPatch(Patch patch)
     {
-        Patches.Add(patch);
+        if (!AnyPatch(patch, Patches))
+            Patches.Add(patch);
     }
 
     public bool HasPatch(Patch patch)
     {
-        return SelectedPatches.Any(p => p.Info.Name == patch.Info.Name);
+        return AnyPatch(patch, SelectedPatches);
     }
 
     public bool ExecuteAllPatches()
@@ -307,8 +132,8 @@ public partial class MainWindow
         {
             if (HasPatch(patch))
             {
-                bool result = patch.Execute();
-                if (result)
+                bool patchSuccess = patch.Execute();
+                if (patchSuccess)
                     success = true;
             }
         }
@@ -316,40 +141,52 @@ public partial class MainWindow
         return success;
     }
 
-    public void ReloadPatchList()
+    public void ReloadPatchList(string searchTerm = "")
     {
-        LoadPatches(true);
         patchList.Items.Clear();
         patchList.BeginUpdate();
         foreach (Patch patch in Patches) {
-            string description = string.IsNullOrEmpty(patch.Info.Description) ? "" : "\n\n" + patch.Info.Description;
-            string author = "\n\nMade by: " + (string.IsNullOrEmpty(patch.Info.Author) ? "N/A" : patch.Info.Author);
-
-            ListViewItem listItem = new ListViewItem(new[] { patch.Info.DisplayName, patch.Info.Version });
-            listItem.Tag = patch;
-            listItem.ToolTipText = $"{patch.Info.DisplayName} ({patch.Info.Name}){description}{author}\nVersion: {patch.Info.Version}";
-
-            if (HasPatch(patch))
-                listItem.Checked = true;
-
-            patchList.Items.Add(listItem);
+            if (string.IsNullOrEmpty(searchTerm) || patch.Info.DisplayName.ToLower().Contains(searchTerm.ToLower()) ||
+                patch.Info.Version.ToLower().Contains(searchTerm.ToLower()) ||
+                patch.Info.Author.ToLower().Contains(searchTerm.ToLower()))
+            {
+                AddPatchToList(patch);
+            }
         }
         patchList.EndUpdate();
+    }
+
+    private void AddPatchToList(Patch patch)
+    {
+        string description = string.IsNullOrEmpty(patch.Info.Description) ? "" : "\n\n" + patch.Info.Description;
+        string author = "\n\nMade by: " + (string.IsNullOrEmpty(patch.Info.Author) ? "N/A" : patch.Info.Author);
+
+        ListViewItem listItem = new ListViewItem(new[] { patch.Info.DisplayName, patch.Info.Version });
+        listItem.Tag = patch;
+        listItem.ToolTipText = $"{patch.Info.DisplayName} ({patch.Info.Name}){description}{author}\nVersion: {patch.Info.Version}";
+
+        if (HasPatch(patch))
+            listItem.Checked = true;
+
+        patchList.Items.Add(listItem);
     }
             
     private void reloadPatches_Click(object sender, EventArgs e)
     {
+        LoadPatches(true);
         ReloadPatchList();
     }
 
     private void searchTextBox_TextChanged(object sender, EventArgs e)
     {
-        ListViewItem? foundItem =
+        // LoadPatches(true);
+        ReloadPatchList(searchTextBox.Text);
+        /*ListViewItem? foundItem =
             patchList.FindItemWithText(searchTextBox.Text, true, 0, true);
         if (foundItem != null)
         {
             patchList.TopItem = foundItem;
-        }
+        }*/
     }
 
     private void patchList_ItemChecked(object sender, ItemCheckedEventArgs e)
@@ -357,13 +194,35 @@ public partial class MainWindow
         Patch? checkedPatch = (Patch)e.Item.Tag;
         if (checkedPatch != null)
         {
-            if (e.Item.Checked && SelectedPatches.Find(p => p.Info.Name == checkedPatch.Info.Name) == null)
+            if (e.Item.Checked && !HasPatch(checkedPatch))
                 SelectedPatches.Add(checkedPatch);
 
-            if (!e.Item.Checked && SelectedPatches.Find(p => p.Info.Name == checkedPatch.Info.Name) != null)
+            if (!e.Item.Checked && HasPatch(checkedPatch))
                 SelectedPatches.Remove(checkedPatch);
 
             SaveSelectedPatches();
         }
+    }
+
+    private void enableAll_Click(object sender, EventArgs e)
+    {
+        foreach (Patch patch in Patches)
+        {
+            if (!HasPatch(patch))
+                SelectedPatches.Add(patch);
+        }
+        SaveSelectedPatches();
+        ReloadPatchList();
+    }
+
+    private void disableAll_Click(object sender, EventArgs e)
+    {
+        foreach (Patch patch in Patches)
+        {
+            if (HasPatch(patch))
+                SelectedPatches.Remove(patch);
+        }
+        SaveSelectedPatches();
+        ReloadPatchList();
     }
 }
